services/voice/requirements.txt
fastapi
uvicorn[standard]
groq
elevenlabs
python-dotenv
python-multipart
